Update formatter.py
degraded whisper to int 8 in line 60

+++ TTS/demos/xtts_ft_demo/utils/formatter.py
@@ -57,7 +57,7 @@ def format_audio_list(
     device = "cuda" if torch.cuda.is_available() else "cpu"
 
     print("Loading Whisper Model!")
-    asr_model = WhisperModel("large-v2", device=device, compute_type="float16")
+    asr_model = WhisperModel("large-v2", device=device, compute_type="int8")
 
     metadata = {"audio_file": [], "text": [], "speaker_name": []}
